+++ src/core/prelude/CHANGELOG.md
@@ -9,6 +9,18 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.70.3 (2021-12-08)
+
+#### :bug: Bug Fix
+
+* Restored the native behavior of `Object.hasOwnProperty`
+
+## v3.70.1 (2021-12-07)
+
+#### :bug: Bug Fix
+
+* Now `Array.union` doesn't spread primitive iterable values `array`
+
 ## v3.67.0 (2021-12-01)
 
 #### :rocket: New Feature