+++ src/core/prelude/array/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.70.1 (2021-12-07)
+
+#### :bug: Bug Fix
+
+* Now `Array.union` doesn't spread primitive iterable values
+
 ## v3.20.0 (2020-07-05)
 
 #### :rocket: New Feature