+++ src/core/promise/abortable/CHANGELOG.md
@@ -9,6 +9,20 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.69.0 (2021-12-03)
+
+#### :boom: Breaking Change
+
+* Removed deprecated API
+
+#### :rocket: New Feature
+
+* Added a new static method `wrapReasonToIgnore`
+
+#### :bug: Bug Fix
+
+* Fixed a bug when a child promise is never resolved because its parent is already rejected
+
 ## v3.60.3 (2021-10-05)
 
 #### :bug: Bug Fix